Added a check to see if the username was in the database before deleting it.

+++ src/command/admin/DeleteAdminUserCommand.java
@@ -44,24 +44,21 @@ public class DeleteAdminUserCommand extends Command {
 
 	@Override
 	public Response execute() {
-		DatabaseAccessor db;
+		DatabaseAccessor db = null;
 		try {
 			db = initDB();
+			if (db.getUsers().contains(username)) {
+				db.deleteUser(username);
+			} else {
+				return new ErrorResponse(HttpStatusCode.BAD_REQUEST,
+						"Deletion of user unsuccessful, user does not exist.");
+			}
 		} catch (SQLException | IOException e) {
 			Debug.log("Deletion of user: " + username + " was unsuccessful, " +
 					"reason: " + e.getMessage());
 			return new ErrorResponse(HttpStatusCode.INTERNAL_SERVER_ERROR,
 					"Deletion of user: " + username + " was unsuccessful due " +
 							"to temporary problems with the database.");
-		}
-
-		try {
-			db.deleteUser(username);
-		} catch (SQLException e) {
-			Debug.log("Deletion of user: " + username + " was unsuccessful, " +
-					"reason: " + e.getMessage());
-			return new ErrorResponse(HttpStatusCode.BAD_REQUEST, "Deletion of "
-					+ "user unsuccessful, user may not exist.");
 		} finally {
 			if (db != null)
 				db.close();
